Create class representing a test-name creator
