+++ modules/http4s-munit/src/main/scala/munit/Http4sMUnitDefaults.scala
@@ -17,41 +17,18 @@
 package munit
 
 import cats.effect.IO
-import cats.syntax.all._
 
 import org.http4s.Request
-import org.http4s.Uri
 
 object Http4sMUnitDefaults {
 
+  @deprecated("Use Http4sMUnitTestNameCreator.default instead", since = "0.16.0")
   def http4sMUnitNameCreator(
       request: Request[IO],
       followingRequests: List[String],
       testOptions: TestOptions,
       config: Http4sMUnitConfig,
       replacements: Seq[(String, String)] = Nil // scalafix:ok
-  ): String = {
-    val clue = followingRequests.+:(testOptions.name).filter(_.nonEmpty) match {
-      case Nil                 => ""
-      case List(head)          => s" ($head)"
-      case List(first, second) => s" ($first and then $second)"
-      case list                => s"${list.init.mkString(" (", ", ", ", and then")} ${list.last})"
-    }
-
-    val context = request.attributes.lookup(RequestContext.key).map(_.asString).filterNot(_.isEmpty())
-
-    val reps = config.repetitions match {
-      case Some(rep) if rep > 1 =>
-        s" - executed $rep times" + config.maxParallel.fold("")(paral => s" with $paral in parallel")
-      case _ => ""
-    }
-
-    val nameWithoutReplacements = s"${request.method.name} -> ${Uri.decode(request.uri.renderString)}" +
-      s"$clue${context.fold("")(" as " + _)}$reps"
-
-    replacements.foldLeft(nameWithoutReplacements) { case (name, (value, replacement)) =>
-      name.replaceAll(value, replacement)
-    }
-  }
+  ): String = Http4sMUnitTestNameCreator.default(replacements).nameFor(request, followingRequests, testOptions, config)
 
 }
